+++ src/ethereum/info/abstractInfo.ts
@@ -66,7 +66,13 @@ const networkInfo: EthereumNetworkInfo = {
   pluginMnemonicKeyName: 'abstractMnemonic',
   pluginRegularKeyName: 'abstractKey',
   evmGasStationUrl: null,
-  networkFees
+  networkFees,
+  decoyAddressConfig: {
+    count: 5,
+    lookbackBlocks: 1000000,
+    minTransactionCount: 10,
+    maxTransactionCount: 100
+  }
 }
 
 const currencyInfo: EdgeCurrencyInfo = {