+++ src/ethereum/info/celoInfo.ts
@@ -91,7 +91,13 @@ const networkInfo: EthereumNetworkInfo = {
   pluginMnemonicKeyName: 'celoMnemonic',
   pluginRegularKeyName: 'celoKey',
   evmGasStationUrl: null,
-  networkFees
+  networkFees,
+  decoyAddressConfig: {
+    count: 5,
+    lookbackBlocks: 1000000,
+    minTransactionCount: 10,
+    maxTransactionCount: 100
+  }
 }
 
 const currencyInfo: EdgeCurrencyInfo = {