+++ src/ethereum/info/ethereumclassicInfo.ts
@@ -112,7 +112,13 @@ const networkInfo: EthereumNetworkInfo = {
   pluginMnemonicKeyName: 'ethereumclassicMnemonic',
   pluginRegularKeyName: 'ethereumclassicKey',
   evmGasStationUrl: null,
-  networkFees
+  networkFees,
+  decoyAddressConfig: {
+    count: 5,
+    lookbackBlocks: 1000000,
+    minTransactionCount: 10,
+    maxTransactionCount: 100
+  }
 }
 
 const currencyInfo: EdgeCurrencyInfo = {