+++ src/ethereum/networkAdapters/FilfoxAdapter.ts
@@ -21,9 +21,10 @@ export interface FilfoxAdapterConfig {
 }
 
 export class FilfoxAdapter extends NetworkAdapter<FilfoxAdapterConfig> {
+  batchMulticastRpc = null
+  broadcast = null
   connect = null
   disconnect = null
-  broadcast = null
   getBaseFeePerGas = null
   multicastRpc = null
   fetchBlockheight = null